refactor(app): split er reducers
